Refs #22518 Moved the generic ObservableDictionary to its own class
